requirements.txt
jupyter==1.0.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.23.2
pandas==1.4.3
pytest==7.1.2
pytorch_lightning==1.2.8
torch==1.12.1+cu116 --find-links https://download.pytorch.org/whl/cu116
torchaudio==0.12.1+cu116 --find-links https://download.pytorch.org/whl/cu116
torchvision==0.13.1+cu116 --find-links https://download.pytorch.org/whl/cu116
torchmetrics==0.9.3